requirements.txt
# Module 1: Data Collection Requirements

# Data handling
pandas>=1.5.0
numpy>=1.21.0

# API requests and environment
requests>=2.28.0
python-dotenv>=0.19.0

# Geospatial data processing
geopandas>=0.12.0
osmnx>=1.2.0
shapely>=1.8.0

# Scientific computing
scipy>=1.9.0

# ML & serialization (for future modules)
scikit-learn>=1.1.0
joblib>=1.2.0
xgboost>=1.6.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.10.0
folium>=0.12.0

# Dashboard (for future modules) 
streamlit>=1.12.0
streamlit_folium>=0.6.0
streamlit_autorefresh>=0.0.1

# Additional utilities
tqdm>=4.64.0  # Progress bars
retry>=0.9.2  # Retry failed API calls